Implement 1D complex convolution.
CONTEXT: We can represent the image formation of an US beam by convolving the response function of the US device with scatterers in the media. To keep track of the phase associated with different spatial areas of the PSF (necessary for the formation of speckle) this must be a complex convolution.

SCOPE: Tensorflow does not have a native complex convolution. This implements a basic 1D complex convolution function.