Implement multi-filter batch convolution.
CONTEXT: The image formation process is modeled by convolving a complex PSF with the scatterer distribution. In order to represent more orthogonal speckle patterns we will have to use multiple frequencies. This is equivalent to using multiple psfs when forming the image.

SCOPE: Previously we could only convolve a single filter at a time with a batch of images.  This CL implements multi-filter convolution.